backend/requirements.txt
Flask
Pillow
numpy